+++ pytsite/auth/__init__.py
@@ -4,7 +4,7 @@
 from . import _error as error, _model as model, _driver as driver
 from ._api import get_current_user, get_user_statuses, get_user, create_user, get_role, get_sign_in_form, find_users, \
     register_driver, user_nickname_rule, sign_in, get_driver, create_role, get_sign_in_url, get_sign_out_url, \
-    verify_password, hash_password, sign_out, sign_in_by_token, get_access_token_info
+    verify_password, hash_password, sign_out, sign_in_by_token, get_access_token_info, sanitize_nickname
 
 __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'
@@ -14,7 +14,7 @@ __license__ = 'MIT'
 def __init():
     """Init wrapper.
     """
-    from pytsite import reg, assetman, odm, events, tpl, lang, router, robots, console, js_api, permission
+    from pytsite import reg, assetman, odm, events, tpl, lang, router, robots, console, permission
     from ._console_command import Passwd as AuthConsoleCommand
     from . import _eh
 
@@ -28,10 +28,6 @@ def __init():
     from . import _model
     odm.register_model('user', _model.User)
     odm.register_model('role', _model.Role)
-
-    # Base auth driver
-    from ._driver.password import Password
-    register_driver(Password())
 
     # Routes
     base_path = reg.get('auth.base_path', '/auth')