+++ pytsite/auth/_driver/__init__.py
@@ -1,9 +1,0 @@
-"""Auth drivers.
-"""
-# Public API
-from .abstract import Abstract
-from .password import Password
-
-__author__ = 'Alexander Shepetko'
-__email__ = 'a@shepetko.com'
-__license__ = 'MIT'