+++ pytsite/auth_driver_password/__init__.py
@@ -1,0 +1,15 @@
+"""PytSite Auth Password Driver Package.
+"""
+
+__author__ = 'Alexander Shepetko'
+__email__ = 'a@shepetko.com'
+__license__ = 'MIT'
+
+
+def _init():
+    from pytsite import auth
+    from . import _driver
+
+    auth.register_driver(_driver.Password())
+
+_init()