+++ pytsite/auth_driver_password/_driver.py
@@ -1,8 +1,6 @@
-"""Password Auth Driver.
+"""PytSite Auth Password Driver.
 """
-from pytsite import form as _form, lang as _lang, widget as _widget, router as _router
-from .. import _api, _error, _model
-from .abstract import Abstract
+from pytsite import auth as _auth, form as _form, router as _router, widget as _widget, lang as _lang
 
 __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'
@@ -36,7 +34,7 @@ class _SignInForm(_form.Form):
         self.get_widget('action-submit').value = _lang.t('pytsite.auth@login')
 
 
-class Password(Abstract):
+class Password(_auth.driver.Abstract):
     """ULogin Driver.
     """
     def get_name(self) -> str:
@@ -57,27 +55,27 @@ class Password(Abstract):
         # TODO
         pass
 
-    def sign_in(self, data: dict) -> _model.User:
+    def sign_in(self, data: dict) -> _auth.model.User:
         """Authenticate user.
         """
         login = data.get('login')
         password = data.get('password')
 
         if not login or not password:
-            raise _error.AuthenticationError('Login or password is not specified.')
+            raise _auth.error.AuthenticationError('Login or password is not specified.')
 
         # Check if the user exists
-        user = _api.get_user(login)
+        user = _auth.get_user(login)
         if not user:
-            raise _error.AuthenticationError(_lang.t('pytsite.auth@authentication_error'))
+            raise _auth.error.AuthenticationError(_lang.t('pytsite.auth@authentication_error'))
 
         # Check password
-        if not _api.verify_password(password, user.password):
-            raise _error.AuthenticationError(_lang.t('pytsite.auth@authentication_error'))
+        if not _auth.verify_password(password, user.password):
+            raise _auth.error.AuthenticationError(_lang.t('pytsite.auth@authentication_error'))
 
         return user
 
-    def sign_out(self, user: _model.User):
+    def sign_out(self, user: _auth.model.User):
         """End user's session.
         """
         pass