+++ pytsite/auth_driver_ulogin/__init__.py
@@ -8,7 +8,7 @@ __license__ = 'MIT'
 
 def _init():
     from pytsite import auth, tpl, assetman
-    from ._auth_driver import ULogin
+    from ._driver import ULogin
 
     auth.register_driver(ULogin())
     tpl.register_package(__name__)