+++ pytsite/google/res/assets/js/platform.js
@@ -1,0 +1,8 @@
+pytsite.google.platform = {
+    ready: false
+};
+
+window.pytsiteGooglePlatformInitCallback = function () {
+    pytsite.google.platform.ready = true;
+    $(window).trigger('pytsite.google.platform.ready');
+};